First GUI build, the pdf2txt GUI
This is NOT a final version of this code, as, in addition to being nonfunctional, I'm also not totally sure what exactly I want this code to do. Ideally this code will allow the user to do ALL of the stuff pdf2txt can do without even needing to download the dependencies first, we'll see.