Shell script section for PDF2TXT install
this shell script will install pdf2txt and update its dependencies when run. as it stands, it uses a non-variable file path and so it will only work on my personal laptop. Next step is to find a way to get it to work on any Linux system by making the file path allow for any directory name to act as user instead of relying on my own computer's username.