Complete wrapper rewrite
The wrapper has been rewritten to now correctly download PDF2TXT and run it all in a self contained program. The wrapper checks for internet connection before running, so that It won't attempt to update if no connection is found, or won't try to install itself if there is no connection to download what it needs to work. The previous version did not contain code to actually run pdf2txt, which this version adds. problems occur with "milvus" that will need to be solved