Connection Check updated
The connection check now checks if github can be accessed instead of google. This resolves a potential issue where the connection check could ping google and get a false negative if google is down or blocked and report that there is no internet connection. Now, instead, the program will ping github and report if a connection can be made. If one can't be, then it doesn't matter if the internet is connected or not, there's no way to download the needed files or updates.