This version is defuncted
This version of the graphical program was never completed, and only builds the graphical interface without any function included. it DOES NOT WORK. It is only being kept in this repo for posterity, and in case I decide to come back to it to make this stuff work in python instead of java.