complete rebuild
The previous version of PDF2TXT built the GUI but had no functionality. This version adds the missing functionality and is designed to work work via argument on the wrapper included here. The wrapper will need a new version so that it works based on arguments instead of by being run through this program, but that can be arranged.