Wrapper with GUI integration
This version of the wrapper is designed explicitly for use with the GUI, using arguments the GUI program generates to run instead of prompting the user for those arguments from the command line. The reason why the GUI uses this wrapper instead of running pdf2txt itself is because the wrapper handles updates and building the file structure, something that Java is unideal for.