CLI version of PDF2TXT wrapper
This version has the same functionality as does the GUI version and GUI combined, minus the benefit of being able to run graphically. For all intents and purposes, this is a complete and working version of PDF2TXT that does in fact severely simplify the process of running pdf2txt.