V 1.1.0
Made better for use from the command line and simplified to remove the update step since it won't be needed with the specific version download the program now uses. Confirmed working from a fresh install. Now referers to a unique file path so that this version won't conflict with the GUI version.