PDF2TXT_GUI 0.4.0
This is the first completely functional version of PDF2TXT graphical. as it stands, running this program alone with the wrappers will do the conversion, and the program itself is capable of downloading PDF2TXT and installing it. next step is to tell the user if/when the installation happens (as currently it is simply a blank screen) and to make it so the conversion no longer requires the absolute file path.